requirements.txt
python-dotenv
slack-bolt
llama-cpp-python